# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "cf-teamspeed API",
    "version": "1.0.0",
    "description": "基于 Cloudflare Realtime SFU + Workers 的私人游戏语音室 API。\n\n**这份文档覆盖完整的核心闭环**（客户端最少需要实现的部分）：配置、鉴权与续期、房间生命周期、WebSocket 握手票、SFU 信令、在线状态。\n\n其余接口（房间管理、管理后台）见 `docs/API.md`，那份是完整清单；本文件刻意保持精简，以便作为「新客户端接入检查表」使用。\n\n鉴权：除 `security: []` 的接口外，一律带 `Authorization: Bearer <token>`。\n\n错误：统一返回 `{ok:false, error, code}`；`code` 是稳定契约，分支判断请用它而不是 `error` 文案。",
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "https://ts.futurvo.cc", "description": "示例部署（请替换成你自己的 Worker 域名）" }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录接口返回的 `token`。有效期 12 小时，随请求滚动续期；服务端会通过响应头 `X-Refreshed-Token` 下发新 token。"
      }
    },
    "headers": {
      "RefreshedToken": {
        "description": "服务端滚动续期下发的新 token。出现时必须覆盖本地保存的 token。",
        "schema": { "type": "string" }
      }
    },
    "schemas": {
      "ApiOk": {
        "type": "object",
        "required": ["ok", "data"],
        "properties": {
          "ok": { "const": true },
          "data": {}
        }
      },
      "ApiError": {
        "type": "object",
        "required": ["ok", "error"],
        "properties": {
          "ok": { "const": false },
          "error": { "type": "string", "description": "给用户看的提示；文案可能变动" },
          "code": { "$ref": "#/components/schemas/ErrorCode" },
          "detail": { "description": "调试信息" }
        }
      },
      "ErrorCode": {
        "type": "string",
        "description": "稳定错误码。客户端分支逻辑必须用它。",
        "enum": [
          "INVALID_KEY",
          "INVALID_BODY",
          "NICKNAME_TAKEN",
          "NICKNAME_RESERVED",
          "NICKNAME_INVALID",
          "UNAUTHORIZED",
          "SESSION_EXPIRED",
          "FORBIDDEN",
          "ROOM_NOT_FOUND",
          "ROOM_FULL",
          "RATE_LIMITED",
          "BANNED",
          "TURNSTILE_FAILED",
          "BAD_TICKET",
          "INTERNAL"
        ]
      },
      "Profile": {
        "type": "object",
        "required": ["uid", "nickname", "role", "avatarId", "avatarUrl"],
        "properties": {
          "uid": { "type": "string", "description": "身份主键；昵称可改，uid 不变" },
          "nickname": { "type": "string" },
          "role": { "type": "string", "enum": ["guest", "admin"] },
          "avatarId": { "type": ["string", "null"], "description": "12 个预设头像之一" },
          "avatarUrl": { "type": ["string", "null"], "description": "自定义头像" }
        }
      },
      "Session": {
        "type": "object",
        "required": ["token", "expiresAt", "sessionExpiresAt", "role", "profile"],
        "properties": {
          "token": { "type": "string", "description": "放进 Authorization: Bearer" },
          "expiresAt": { "type": "integer", "description": "token 过期时间（Unix 秒）" },
          "sessionExpiresAt": { "type": "integer", "description": "会话绝对上限（Unix 秒），到点必须重新登录" },
          "role": { "type": "string", "enum": ["guest", "admin"] },
          "profile": { "$ref": "#/components/schemas/Profile" }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "未登录 / token 无效 / 会话到顶（`code` 分别为 `UNAUTHORIZED`、`SESSION_EXPIRED`）",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
        }
      },
      "RateLimited": {
        "description": "请求过于频繁，按 `Retry-After` 头退避",
        "headers": { "Retry-After": { "schema": { "type": "integer" }, "description": "建议等待秒数" } },
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
        }
      }
    }
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "用 key + 昵称换会话 token",
        "description": "昵称全局唯一，已被占用返回 `NICKNAME_TAKEN`。管理员 key 登录需要 `turnstileToken`（见 `config.adminLoginTurnstile`）。",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["key", "nickname"],
                "properties": {
                  "key": { "type": "string" },
                  "nickname": { "type": "string", "minLength": 2, "maxLength": 16 },
                  "avatarId": { "type": ["string", "null"] },
                  "turnstileToken": { "type": ["string", "null"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "headers": {
              "X-Refreshed-Token": { "$ref": "#/components/headers/RefreshedToken" },
              "X-Auth-Method": { "schema": { "type": "string" }, "description": "本次用了哪种凭据" }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiOk" },
                    { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Session" } } }
                  ]
                }
              }
            }
          },
          "401": { "description": "key 无效（`INVALID_KEY`）" },
          "403": { "description": "人机验证失败（`TURNSTILE_FAILED`）或被封禁（`BANNED`）" },
          "409": { "description": "昵称不可用（`NICKNAME_TAKEN` / `NICKNAME_RESERVED` / `NICKNAME_INVALID`）" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "summary": "会话续期",
        "description": "其实任何鉴权请求都会滚动续期；这个接口给「整晚只挂 WebSocket」的客户端一个确定的续期时机。会话到绝对上限时返回 `SESSION_EXPIRED`，此时必须重新登录。",
        "responses": {
          "200": {
            "description": "续期成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiOk" },
                    { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Session" } } }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["auth"],
        "summary": "当前会话与资料",
        "responses": {
          "200": {
            "description": "当前资料",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiOk" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "profile": { "$ref": "#/components/schemas/Profile" },
                            "authMethod": { "type": "string", "enum": ["cookie", "bearer"] },
                            "expiresAt": { "type": ["integer", "null"] },
                            "sessionExpiresAt": { "type": "integer" }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "patch": {
        "tags": ["auth"],
        "summary": "修改昵称 / 头像",
        "description": "改名会原地同步到在线名册与所在房间，别人立刻看到新昵称；uid 不变。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": { "type": "string" },
                  "avatarId": { "type": ["string", "null"] },
                  "avatarDataUrl": {
                    "type": ["string", "null"],
                    "description": "data:image/(png|jpeg|webp);base64,…  ≤200KB"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "更新成功，返回新资料与新 token" },
          "400": { "description": "昵称不可用（`NICKNAME_TAKEN` 等）" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
...
}